fix(2-dbt): cast Fanbasis money to NUMERIC, not FLOAT64
The staging.md rule explicitly mandates `cast(amount as numeric)` for
money fields. stg_fanbasis__transactions and stg_fanbasis__refunds were
violating this with float64 casts; PR #92's parity test absorbed the
drift via $0.01 tolerance.

Changes:
- stg_fanbasis__{transactions,refunds}: float64 → numeric on all amount
  columns (gross/net/fee, refund/net/fee/total).
- fct_payments Stripe arm: `amount_minor / 100.0` (FLOAT64) → `cast as
  numeric / 100` (NUMERIC) so the UNION ALL with the now-NUMERIC
  Fanbasis arm doesn't fail with type mismatch.
- release_gate_revenue_detail: pct-delta thresholds cast as `numeric
  '0.05'` / `numeric '0.10'` so the gate arithmetic stays in NUMERIC
  end-to-end (avoids implicit FLOAT64 coercion at comparison that would
  re-introduce the precision drift the rest of the PR removes).
- revenue_detail_refunds_parity: comment-only update; tolerance kept at
  $0.01 as a precision floor (now effectively exact).

Verified: `dbt parse` clean; ce-adversarial-reviewer Go-with-changes
(release_gate hardening incorporated). Persistent tables (fct_payments,
fct_refunds, revenue_detail) absorb the type flip via dbt's default
table drop+create. No downstream views select from these — so no schema
ALTER hazard. dabi BI rendering of NUMERIC vs FLOAT64 is the only
unverified surface; will surface at first dabi run post-merge.

Closes the "Float64-for-money tech debt (Fanbasis)" thread.

+++ 2-dbt/models/staging/fanbasis/stg_fanbasis__refunds.sql
@@ -63,10 +63,10 @@ parsed as (
         -- silently misattribute the contact otherwise).
         parent_payment_id                                             as payment_id,
 
-        cast(json_value(refund_json, '$.amount') as float64)          as refund_amount,
-        cast(json_value(refund_json, '$.amount_gross') as float64)    as refund_amount_net,
-        cast(json_value(refund_json, '$.fee') as float64)             as refund_fee,
-        cast(json_value(refund_json, '$.refund_cost') as float64)     as refund_total_cost,
+        cast(json_value(refund_json, '$.amount') as numeric)          as refund_amount,
+        cast(json_value(refund_json, '$.amount_gross') as numeric)    as refund_amount_net,
+        cast(json_value(refund_json, '$.fee') as numeric)             as refund_fee,
+        cast(json_value(refund_json, '$.refund_cost') as numeric)     as refund_total_cost,
         'usd'                                                         as currency,
 
         cast(json_value(refund_json, '$.created_at') as timestamp)    as refunded_at,
